Alterando tipo de retorno

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -18,7 +18,7 @@ public class Calculadora {
 		int soma = soma(a, b);
 		int subtracao = subtracao(a, b);
 		int multiplicacao = multiplicacao(a,b);
-		int divisao = divisao(a, b);
+		double divisao = divisao(a, b);
 		
 		System.out.println(soma);
 		System.out.println(subtracao);
@@ -36,7 +36,7 @@ public class Calculadora {
 		return a - b;
 	}
 
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, int b) {
 		return a / b;
 	}
